File recovery (whew!). Also made changes to make `PARMAETERS` and `METHODS` be changed globally. #13
