+++ param.config
@@ -23,10 +23,10 @@ default_catalog: SDSS-Jan-25-2017-merged.fits
 # Default output file prexfix (includes x and y labels automatically)
 default_prefix: TEST
 
-# Default covariate variable ('independent' variable)
-default_x: lambda
+# Default covariate variable (independent variable)
+default_x: None
 
-# Default response variable ('dependent' variable)
+# Default response variable (dependent variable)
 default_y: tr2500
 
 # Column name conventions:
